Rename inputs operator file
