+++ src/operators/index.ts
@@ -2,4 +2,4 @@ export * from "./frames";
 export * from "./stocks";
 export * from "./player";
 export * from "./pausable";
-export * from "./throttleInputButtons";
+export * from "./inputs";
