+++ src/operators/inputs.spec.ts
@@ -1,7 +1,7 @@
-import {TestScheduler} from "rxjs/testing";
+import { TestScheduler } from "rxjs/testing";
 import { from } from "rxjs";
 import { InputButtonCombo } from "..";
-import { throttleInputButtons } from "./throttleInputButtons";
+import { throttleInputButtons } from "./inputs";
 
 describe("throttleInputButtons operator", () => {
   let testScheduler: TestScheduler;
@@ -17,7 +17,7 @@ describe("throttleInputButtons operator", () => {
     const a: InputButtonCombo = {
       frame: 30,
       playerIndex: 0,
-      combo:  ["X"],
+      combo: ["X"],
       duration: 1,
     };
     const a2: InputButtonCombo = {
@@ -27,24 +27,24 @@ describe("throttleInputButtons operator", () => {
     const b: InputButtonCombo = {
       ...a,
       frame: 78,
-    }
+    };
     const b2: InputButtonCombo = {
       ...a,
       frame: 112,
-    }
+    };
     const c: InputButtonCombo = {
       ...a,
       frame: 113,
-    }
+    };
 
-    testScheduler.run(({expectObservable}) => {
+    testScheduler.run(({ expectObservable }) => {
       // We expect 3 values
       const expectedMarble = "a  b  (c|)";
       const expectedValues = { a, b, c };
       const source$ = from([a, a2, b, b2, c]).pipe(
-        throttleInputButtons(35),  // Throttle for 35 frames
-      )
+        throttleInputButtons(35), // Throttle for 35 frames
+      );
       expectObservable(source$).toBe(expectedMarble, expectedValues);
     });
   });
-})
+});
