+++ src/operators/inputs.ts
